:truck: Rename folder
- rename `interfaces` to `types`
